requirements.txt
# streamlit
# PyPDF2
# faiss-cpu
# torch
# sentence-transformers
# google-generativeai
# numpy
# python-dotenv
# pip install streamlit supabase sentence-transformers PyPDF2 faiss-cpu beautifulsoup4 requests python-dotenv google-generativeai
# streamlit
# sentence-transformers
# supabase
# google-generativeai
# python-dotenv
# selenium
# webdriver-manager
# PyMuPDF
# nest_asyncio
# 
# SUPABASE_URL=https://cicwnaertxmwixtrttid.supabase.co
# SUPABASE_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImNpY3duYWVydHhtd2l4dHJ0dGlkIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDQxMTU2NDksImV4cCI6MjA1OTY5MTY0OX0._9zpsF8AV607vl_rRMtbbyiT9fmv1o_j_qBFRVMBY2Y
# API_KEY=AIzaSyC87pJSv_ocfWQoKgdSP7aeNT3kYxSDguk
# pip install -r requirements.txt
# streamlit run app.py
streamlit
PyMuPDF
nest_asyncio
python-dotenv
sentence-transformers
supabase
google-generativeai
selenium
webdriver-manager


